Fix VFS file struct leak, vnode use-after-free, and user-space bugs
VFS: fut_vfs_close() and close_fd_in_task() now only perform vnode
cleanup and free the file struct when refcount reaches 0. Previously,
every close called vnode->ops->close and fut_vnode_unref regardless of
refcount, causing use-after-free on dup'd file descriptors and leaking
the struct fut_file on every file close.

User-space: Add integer overflow check to calloc(), fix memory leak in
closedir() error paths, fix size_t-to-int truncation in copy_iov_data().

+++ src/user/libfutura/socket_unix.c
@@ -323,7 +323,7 @@ static void free_packet(struct unix_packet *pkt) {
     free(pkt);
 }
 
-static int copy_iov_data(uint8_t *dst, const struct iovec *iov, size_t iovlen) {
+static size_t copy_iov_data(uint8_t *dst, const struct iovec *iov, size_t iovlen) {
     size_t offset = 0;
     for (size_t i = 0; i < iovlen; ++i) {
         if (!iov[i].iov_base || iov[i].iov_len == 0) {
@@ -332,7 +332,7 @@ static int copy_iov_data(uint8_t *dst, const struct iovec *iov, size_t iovlen) {
         memcpy(dst + offset, iov[i].iov_base, iov[i].iov_len);
         offset += iov[i].iov_len;
     }
-    return (int)offset;
+    return offset;
 }
 
 static size_t iov_total_length(const struct iovec *iov, size_t iovlen) {
